Move RaceExtensions to own file, follow convention.
