+++ W3ChampionsStatisticService/CommonValueObjects/Race.cs
@@ -20,21 +20,4 @@ namespace W3ChampionsStatisticService.CommonValueObjects
         UD = 3,
         NE = 4,
     }
-
-    public static class Extensions
-    {
-        static Dictionary<RaceId, Race> raceIdMap = new Dictionary<RaceId, Race>() {
-            { RaceId.RnD, Race.RnD },
-            { RaceId.HU, Race.HU },
-            { RaceId.OC, Race.OC },
-            { RaceId.NE, Race.NE },
-            { RaceId.UD, Race.UD }
-        };
-
-        public static Race FromRaceId(this Race race, RaceId raceId)
-        {
-            return raceIdMap[raceId];
-        }
-    }
-
 }
